MAINT: refactoring cli app
