+++ enigma-principal/app/src/cli/mod.rs
@@ -1,1 +1,2 @@
 pub mod options;
+pub mod app;
